requirements.txt
astroid==2.3.2
certifi==2019.9.11
chainer==6.5.0
chainercv==0.13.1
chardet==3.0.4
Click==7.0
filelock==3.0.12
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.3
pandas==0.25.2
Pillow==6.2.1
protobuf==3.10.0
pylint==2.4.3
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.10
typed-ast==1.4.0
typing==3.6.6
typing-extensions==3.6.6
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
